REGRESSION(r209653) Crash in CallFrameShuffler::snapshot() https://bugs.webkit.org/show_bug.cgi?id=165728
Reviewed by Filip Pizlo.

JSTests:

New regression test.

* stress/regress-165728.js: Added.
(sum1):
(sum2):
(tailCaller):
(test):

Source/JavaScriptCore:

It can be the case that a JSValueReg's CachedRecovery is the source for mutliple
GPRs. We only store the CachedRecovery in one slot of m_newRegisters to simplify
the recovery process. This is also done for the case where the recovery source
and destination are the same GPR.

In light of this change, snapshot needs to be taught that one CacheRecovery is
the source for multiple registers.  This is done by using a two step process.
First find all the argument CachedRecovery's and create a vector mapping all of
the target GPRs and the source recovery.  Then use that vector to get the
recovery for each register.

* jit/CallFrameShuffler.h:
(JSC::CallFrameShuffler::snapshot):


Identifier: 183329@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209673 268f45cc-cd09-0410-ab3c-d52691b4dbfc