text-align: justify and word-spacing combine to overflow column https://bugs.webkit.org/show_bug.cgi?id=165796 rdar://problem/29672168
Reviewed by Myles C. Maxfield.

Source/WebCore:

The value of the word-spacing property is not applied on the inline-tree boxes (normal line layout)
with kerning enabled. It causes hittest failures and selection/rendering glitches.
We normally use the pre-computed 'WordMeasurements' values to figure out the run widths. However
with kerning on, in some cases we need to re-measure a certain part of the text by calling RenderText::width().
This function omits the word-spacing value for leading whitespace, so when this text fragment starts with a whitespace,
(while it's not at the beginning of the run) we have to manually add the word-spacing value back.
r146087 added this logic with the restriction of applying it only to whitespace-only runs.

Test: fast/css/word-spacing-with-normal-layout.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::setLogicalWidthForTextRun):

LayoutTests:

These are progressions.

* fast/css/word-spacing-with-normal-layout-expected.html: Added.
* fast/css/word-spacing-with-normal-layout.html: Added.
* platform/mac/css1/text_properties/word_spacing-expected.txt:
* platform/mac/fast/css/word-space-extra-expected.txt:
* platform/mac/fast/text/whitespace/span-in-word-space-causes-overflow-expected.txt:
* platform/mac/fast/text/word-space-expected.txt:
* platform/mac/svg/custom/svg-fonts-in-html-expected.txt:
* platform/mac/svg/custom/svg-fonts-word-spacing-expected.txt:


Identifier: 183544@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209910 268f45cc-cd09-0410-ab3c-d52691b4dbfc