CellState should have members with accurate names https://bugs.webkit.org/show_bug.cgi?id=165969
Reviewed by Mark Lam.

This once again renames the members in CellState. I wanted to convey the following
pieces of information in the names:

- What does the state mean for Generational GC?
- What does the state mean for Concurrent GC?
- Does the state guarantee what it means, or is there some contingency?

The names I came up with are:

PossiblyOldOrBlack: An object in this state may be old, or may be black, depending on
    other things. If the mark bit is set then the object is either black or being
    blackened as we speak. It's going to survive the GC, so it will be old, but may be
    new now. In between GCs, objects in this state are definitely old. If the mark bit
    is not set, then the object is actually old and white.

DefinitelyNewAndWhite: The object was just allocated so it is white (not marked) and
    new.

DefinitelyGrey: The object is definitely grey - it will be rescanned in the future. It
    may be new or old depending on other things.

* heap/CellState.h:
* heap/Heap.cpp:
(JSC::Heap::addToRememberedSet):
(JSC::Heap::writeBarrierSlowPath):
* heap/SlotVisitor.cpp:
(JSC::SlotVisitor::appendJSCellOrAuxiliary):
(JSC::SlotVisitor::setMarkedAndAppendToMarkStack):
(JSC::SlotVisitor::appendToMarkStack):
(JSC::SlotVisitor::visitChildren):
* runtime/JSCellInlines.h:
(JSC::JSCell::JSCell):
* runtime/StructureIDBlob.h:
(JSC::StructureIDBlob::StructureIDBlob):



Identifier: 183584@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209954 268f45cc-cd09-0410-ab3c-d52691b4dbfc