[EME] Remove MediaKeySessionType::persistent-usage-record https://bugs.webkit.org/show_bug.cgi?id=166012
Reviewed by Xabier Rodriguez-Calvar.

Remove the "persistent-usage-record" value from the MediaKeySessionType.
This was removed from the spec as an at-risk feature.
https://github.com/w3c/encrypted-media/issues/353

No non-imported tests need to be updated. This is still present in the
tests imported from the W3C's web-platform-tests repository, but the
tests haven't yet been updated upstream.

* Modules/encryptedmedia/CDM.cpp:
(WebCore::CDM::isPersistentType):
* Modules/encryptedmedia/MediaKeySessionType.h:
* Modules/encryptedmedia/MediaKeySessionType.idl:


Identifier: 183612@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209983 268f45cc-cd09-0410-ab3c-d52691b4dbfc