[Win] jsc.exe sometimes never exits. https://bugs.webkit.org/show_bug.cgi?id=158073
Reviewed by Darin Adler.

On Windows the thread specific destructor is also called when the main thread is exiting.
This may lead to the main thread waiting forever for the machine thread lock when exiting,
if the sampling profiler thread was terminated by the system while holding the machine
thread lock.

* heap/MachineStackMarker.cpp:
(JSC::MachineThreads::removeThread):


Identifier: 183797@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210237 268f45cc-cd09-0410-ab3c-d52691b4dbfc