Web Inspector: opening Test.html in a normal browser window doesn't log errors to console https://bugs.webkit.org/show_bug.cgi?id=166570
Reviewed by Joseph Pecoraro.

Early syntax errors in the test harness should be logged to the page console since they are easier to
debug in a normal browser using Web Inspector. But, the checks to revert to normal console don't work.

* UserInterface/Test/FrontendTestHarness.js:
(FrontendTestHarness.prototype.reportUncaughtException):
Add a helper to encapsulate the meaning of checking this._shouldResendResults.
This flag is always true until the test page injects its initializers into the
inspector page, which will never happen when we view Test.html outside of the test harness.


Identifier: 183806@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210249 268f45cc-cd09-0410-ab3c-d52691b4dbfc